refactor(cef): split launcher_ipc.rs and app.rs
Tier 2 item from docs/reports/REPORT_LARGE_FILE_MODULARIZATION_SCAN_2026_07_22.md.

launcher_ipc.rs: platform-triplicated connection setup (Windows/Unix/stub,
inherently large) vs. a family of 21 uniform, stateless report_* functions
that just build a Command and send on COMMAND_TX. Extracted the reporters
into launcher_ipc/reporters.rs, re-exported from mod.rs so the ~40 existing
crate::launcher_ipc::report_* call sites needed zero changes.

app.rs: four wrap_*! CEF-delegate macro blocks (legitimately large FFI
boilerplate, kept together) vs. standalone platform-utility functions with
no dependency on the delegates (GPU-tier detection, monitor/DPI geometry,
Linux window-properties override). Extracted app/{gpu,monitor,
window_settings}.rs, re-exported from app/mod.rs so external
crate::app::* call sites needed zero changes.

Pure relocation, no logic changes. cargo build -p agentmux-cef clean;
cargo test -p agentmux-cef launcher_ipc:: app:: — counts match before/after
(3/3 and 0/0 respectively — app.rs had no tests originally either).

<!-- agentmux:agent_id=agentx -->
